Add files via upload
I made a weight and workout program to track me and a coworkers weight and workout. It save to a webpage and opens up to view the saves. I eventually would like to make it where you can edit the saves. Since I have only be learning python on my own. This is the best I have for now.